Move ring datastructure from sorted list to an AVL tree
The ring data structure based on a combination of a hash and a sorted
list are less than ideal for a large number of nodes in the ring.

The change uses an AVL tree instead. The AVL Tree implementation is
augmented so that given a key, you can find the a key in the AVL tree
that would sequentially follow the provided key.